+++ src/FarNet.Redis/Commands/BaseDBCommand.cs
@@ -1,0 +1,8 @@
+﻿using StackExchange.Redis;
+
+namespace FarNet.Redis.Commands;
+
+public abstract class BaseDBCommand : AnyCommand
+{
+    public required IDatabase Database { get; init; }
+}