+++ src/FarNet.Redis/Commands/BaseKeyCommand.cs
@@ -1,0 +1,8 @@
+﻿using StackExchange.Redis;
+
+namespace FarNet.Redis.Commands;
+
+public abstract class BaseKeyCommand : BaseDBCommand
+{
+    public required RedisKey Key { get; init; }
+}