+++ src/FarNet.Redis/Commands/WaitString.cs
@@ -1,0 +1,39 @@
+﻿using System;
+using System.Diagnostics;
+using System.Threading;
+
+namespace FarNet.Redis.Commands;
+
+public sealed class WaitString : BaseKeyCommand
+{
+    // required
+    public required TimeSpan Delay { get; init; }
+    public required TimeSpan Timeout { get; init; }
+
+    // result
+    public string? Result { get; private set; }
+
+    protected override void Validate()
+    {
+        AssertDelayTimeout(Delay, Timeout);
+    }
+
+    protected override void Execute()
+    {
+        var sw = Stopwatch.StartNew();
+        while (true)
+        {
+            var res = Database.StringGet(Key);
+            if (!res.IsNull)
+            {
+                Result = res;
+                return;
+            }
+
+            if (sw.Elapsed >= Timeout)
+                return;
+
+            Thread.Sleep(Delay);
+        }
+    }
+}