+++ tests/About.ps1
@@ -1,4 +1,4 @@
 ﻿Set-StrictMode -Version 3
 
 Import-Module FarNet.Redis
-$Global:db = Open-Redis 127.0.0.1:3278
+$db = Open-Redis 127.0.0.1:3278